feat(core): refactor theme service, support antd theme
