+++ packages/mana-core/src/theme/basic/index.ts
@@ -1,3 +1,2 @@
 export * from './variable-registry';
-export * from './variable-application';
 export * from './variable-protocol';
