+++ packages/mana-core/src/theme/color/index.ts
@@ -1,3 +1,2 @@
 export * from './color-registry';
-export * from './color-application';
 export * from './color-protocol';
