+++ packages/mana-core/src/theme/protocol.ts
@@ -1,0 +1,16 @@
+export type CssVariable = {
+  name: string;
+  value: string;
+};
+
+export interface VariableDefinition<T = any> {
+  id: string;
+  defaults?:
+    | {
+        light: T;
+        dark: T;
+        [key: string]: T;
+      }
+    | undefined;
+  description: string;
+}
